glbaguni-backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "glbaguni-backend"
version = "3.1.0"
description = "AI-powered RSS summarization backend service"
authors = [{name = "Glbaguni Team"}]
dependencies = [
    "openai>=1.23.0",
    "httpx>=0.27.0", 
    "python-dotenv==1.0.0",
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "feedparser==6.0.10",
    "beautifulsoup4==4.12.2",
    "requests==2.31.0",
    "pydantic==2.5.0",
    "sqlalchemy==2.0.23",
]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
src_paths = ["backend", "tests"]

[tool.pylint]
[tool.pylint.messages_control]
disable = [
    "missing-docstring",
    "too-few-public-methods", 
    "too-many-arguments",
    "too-many-locals",
    "duplicate-code",
    "import-error"
]

[tool.pylint.format]
max-line-length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = [
    "tests/",
    "__pycache__/",
] 